+++ DesignToolbox/DesignToolbox/Pages/Components/Navigations/ToolBar/ToolBarTop/ToolBarTopConfigurationView.swift
@@ -68,7 +68,7 @@ final class ToolBarTopConfigurationModel: ToolBarConfigurationModel {
         isTrailingEnabled = true
         isTrailingEmphasized = false
 
-        ios26ButtonStyle = .proiminent
+        ios26ButtonStyle = .prominent
     }
 
     deinit {}